% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{motis_connect}
\alias{motis_connect}
\title{Connect to a local MOTIS IPC endpoint}
\usage{
motis_connect(
  address = getOption("motis.nng.address", motis_default_address())
)
}
\arguments{
\item{address}{NNG IPC address, e.g. "ipc:///tmp/motis-ipc.sock".}
}
\value{
Invisibly returns the underlying nanonext socket.
}
\description{
Connect to a local MOTIS IPC endpoint
}
