% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{motis_default_address}
\alias{motis_default_address}
\title{Default IPC address helper}
\usage{
motis_default_address()
}
\value{
Default IPC address matching the MOTIS server default.
}
\description{
Default IPC address helper
}
\keyword{internal}
